Add scripts and data
