+++ src/llm/prompt_templates.py
@@ -121,12 +121,12 @@ SUMMARY_QUESTIONS = {
             
             Пошаговые инструкции:
             1. Используйте только информацию, предоставленную вам в интервью и в выявленных категориях проблем, не придумывайте ничего самостоятельно.
-            2. Сделайте очень краткое резюме всех разделов "Шаг:1. ОСНОВНЫЕ РАЗДЕЛЫ". Обобщите данные и дайте ответ в формате json, где ключом будет "step_01", а значениями будут обобщенные данные по разделу "Шаг:1. ОСНОВНЫЕ РАЗДЕЛЫ". 
-            3. Сделайте очень резюме всех разделов "Шаг:2. КОДИРОВАНИЕ ИНТЕРВЬЮ". Обобщите данные и дайте ответ в формате json, где ключом будет "step_02", а значениями будут обобщенные данные по разделу "Шаг:2. КОДИРОВАНИЕ ИНТЕРВЬЮ". 
-            4. Сделайте очень резюме всех разделов "Шаг:3. КАТЕГОРИИ КОДОВ В ИНТЕРВЬЮ". Обобщите данные и дайте ответ в формате json, где ключом будет "step_03", а значениями будут обобщенные данные по разделу "Шаг:3. КАТЕГОРИИ КОДОВ В ИНТЕРВЬЮ".
-            5. Сделайте очень резюме всех разделов "Шаг:4. ПРОБЛЕМЫ ОБОЗНАЧЕННЫЕ РЕСПОНДЕНТОМ В ИНТЕРВЬЮ". Обобщите данные и дайте ответ в формате json, где ключом будет "step_04", а значениями будут обобщенные данные по разделу "Шаг:4. ПРОБЛЕМЫ ОБОЗНАЧЕННЫЕ РЕСПОНДЕНТОМ В ИНТЕРВЬЮ".
-            6. Сделайте очень резюме всех разделов "Шаг:5. СГРУППИРОВАННЫЕ ПРОБЛЕМЫ". Обобщите данные и дайте ответ в формате json, где ключом будет "step_05", а значениями будут обобщенные данные по разделу "Шаг:5. СГРУППИРОВАННЫЕ ПРОБЛЕМЫ".
-            7. Сделайте очень резюме всех разделов "Шаг:6. ИТОГОВЫЕ ВЫВОДЫ ПО ИНТЕРВЬЮ". Обобщите данные и дайте ответ в формате json, где ключом будет "step_06", а значениями будут обобщенные данные по разделу "Шаг:6. ИТОГОВЫЕ ВЫВОДЫ ПО ИНТЕРВЬЮ".
+            2. Сделайте резюме всех разделов "Шаг:1. ОСНОВНЫЕ РАЗДЕЛЫ". Обобщите данные и дайте ответ в формате json, где ключом будет "step_01", а значениями будут обобщенные данные по разделу "Шаг:1. ОСНОВНЫЕ РАЗДЕЛЫ". 
+            3. Сделайте резюме всех разделов "Шаг:2. КОДИРОВАНИЕ ИНТЕРВЬЮ". Обобщите данные и дайте ответ в формате json, где ключом будет "step_02", а значениями будут обобщенные данные по разделу "Шаг:2. КОДИРОВАНИЕ ИНТЕРВЬЮ". 
+            4. Сделайте резюме всех разделов "Шаг:3. КАТЕГОРИИ КОДОВ В ИНТЕРВЬЮ". Обобщите данные и дайте ответ в формате json, где ключом будет "step_03", а значениями будут обобщенные данные по разделу "Шаг:3. КАТЕГОРИИ КОДОВ В ИНТЕРВЬЮ".
+            5. Сделайте резюме всех разделов "Шаг:4. ПРОБЛЕМЫ ОБОЗНАЧЕННЫЕ РЕСПОНДЕНТОМ В ИНТЕРВЬЮ". Обобщите данные и дайте ответ в формате json, где ключом будет "step_04", а значениями будут обобщенные данные по разделу "Шаг:4. ПРОБЛЕМЫ ОБОЗНАЧЕННЫЕ РЕСПОНДЕНТОМ В ИНТЕРВЬЮ".
+            6. Сделайте резюме всех разделов "Шаг:5. СГРУППИРОВАННЫЕ ПРОБЛЕМЫ". Обобщите данные и дайте ответ в формате json, где ключом будет "step_05", а значениями будут обобщенные данные по разделу "Шаг:5. СГРУППИРОВАННЫЕ ПРОБЛЕМЫ".
+            7. Сделайте  резюме всех разделов "Шаг:6. ИТОГОВЫЕ ВЫВОДЫ ПО ИНТЕРВЬЮ". Обобщите данные и дайте ответ в формате json, где ключом будет "step_06", а значениями будут обобщенные данные по разделу "Шаг:6. ИТОГОВЫЕ ВЫВОДЫ ПО ИНТЕРВЬЮ".
             8. В итоговом ответе должны быть данные только в json формате, это важно. 
             
             Данные для анализа: <interview>{raw_interview}</interview>
